requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
rich==13.9.4
SQLAlchemy==2.0.37
typing_extensions==4.12.2
